requirements.txt
python-dotenv==0.20.0
keyboard==0.13.5
PyAutoGUI==0.9.53